requirements.txt
firecrawl-py>=4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.32.0
